rag-course/01-faiss-only/requirements.txt

sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.2
